Moved server package to contrib/pkg, fixed imports.
